requirements.txt
torch
torchvision
pandas
Pillow
matplotlib
scikit-learn
